Optimised scopus now has multiple search types for selectors
New equality check for ids
Added potential for data merge in post process
Abstracted abstract selector and data types
